requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytube==15.0.0
colorama==0.4.6
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2

